@@ -2,7 +2,7 @@ package se.gu.spraakbanken.fcs.endpoint.korp.data.json.pojo.statistics;
 
 import java.util.HashMap;
 import java.util.Map;
-import javax.annotation.Generated;
+import jakarta.annotation.Generated;
 import com.fasterxml.jackson.annotation.JsonAnyGetter;
 import com.fasterxml.jackson.annotation.JsonAnySetter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
